+++ elements/role/role.router.ts
@@ -5,14 +5,12 @@ import ConnectedRequest from "../../globalTypes/ConnectedRequest";
 import PaginationResponse from "../../globalTypes/PaginationResponse";
 import ResponseDto from "../../globalTypes/ResponseDto";
 import protectMiddleware from "../../middleware/protectMiddleware";
-import superAdminProtectMiddleware from "../../middleware/superAdminProtectMiddleware";
-import { SuperRole } from "../user/user.model";
 import RoleCreateCommand from "./dto/RoleCreateCommand";
 import RoleReadDto, { toReadDto } from "./dto/RoleReadDto";
 import RolesGetCommand from "./dto/RolesGetCommand";
 import RolesSearchCommand from "./dto/RolesSearchCommand";
 import RoleUpdateCommand from "./dto/RoleUpdateCommand";
-import { IRole } from "./role.model";
+import { IRole, Permission } from "./role.model";
 import roleService from "./role.service";
 
 const router = Router();
@@ -20,15 +18,16 @@ const router = Router();
 router.post(
   "/",
   protectMiddleware,
-  superAdminProtectMiddleware,
   async (
     req: ConnectedRequest<any, any, RoleCreateCommand, any>,
     res: Response<ResponseDto<RoleReadDto>>
   ) => {
     const command: RoleCreateCommand = req.body;
-    if (req.user.superRole !== SuperRole.SuperAdmin) {
-      throw new Error("Unauthorized to create field");
-    }
+    roleService.checkPermission({
+      user: req.user,
+      permission: Permission.CreateRole,
+    });
+
     const field: IRole = await roleService.createRole(command);
 
     return res.status(200).send({
@@ -41,11 +40,15 @@ router.post(
 router.put(
   "/",
   protectMiddleware,
-  superAdminProtectMiddleware,
   async (
     req: ConnectedRequest<any, any, RoleUpdateCommand, any>,
     res: Response<ResponseDto<RoleReadDto>>
   ) => {
+    roleService.checkPermission({
+      user: req.user,
+      permission: Permission.UpdateRole,
+    });
+
     const command: RoleUpdateCommand = req.body;
 
     const role: IRole = await roleService.updateRole(command);
@@ -59,10 +62,16 @@ router.put(
 
 router.post(
   "/getRoles",
+  protectMiddleware,
   async (
     req: ConnectedRequest<any, any, RolesGetCommand, any>,
     res: Response<ResponseDto<PaginationResponse<RoleReadDto>>>
   ) => {
+    roleService.checkPermission({
+      user: req.user,
+      permission: Permission.ReadRole,
+    });
+
     const command: RolesGetCommand = req.body;
     const { roles, total } = await roleService.getRoles(command);
 
@@ -79,11 +88,15 @@ router.post(
 router.delete(
   "/",
   protectMiddleware,
-  superAdminProtectMiddleware,
   async (
     req: ConnectedRequest<any, any, mongoose.ObjectId[], any>,
     res: Response<ResponseDto<void>>
   ) => {
+    roleService.checkPermission({
+      user: req.user,
+      permission: Permission.DeleteRole,
+    });
+
     const rolesIds: mongoose.ObjectId[] = req.body;
     await roleService.deleteRoles(rolesIds);
 
@@ -96,10 +109,16 @@ router.delete(
 
 router.post(
   "/search",
+  protectMiddleware,
   async (
     req: ConnectedRequest<any, any, RolesSearchCommand, any>,
     res: Response<ResponseDto<PaginationResponse<RoleReadDto>>>
   ) => {
+    roleService.checkPermission({
+      user: req.user,
+      permission: Permission.ReadRole,
+    });
+
     const command: RolesSearchCommand = req.body;
 
     const { roles, total } = await roleService.search(command);
